Move `UserConfig` to its own file
